+++ History.md
@@ -1,3 +1,10 @@
+3.2.0 / 2022-06-09
+==================
+  **updates**
+  * Update auto number grab tools
+  * Added fake numberplate detector
+
+
 3.1.0 / 2022-03-28
 ==================
   **updates**